+++ R/signal.R
@@ -55,17 +55,15 @@ lifecycle_cnd_data <- function(cnd) {
     signaller <- "signal_superseded"
   } else {
     class <- paste(class(cnd), collapse = "/")
-    abort(glue::glue(
-      "Internal error: Unsupported class `{class}` in `lifecycle_cnd_data()`."
-    ))
+    lifecycle_abort("Unsupported class `{class}` in `lifecycle_cnd_data()`.")
   }
 
-  what <- feature_spec(cnd$what, signaller = signaller)
+  what <- spec(cnd$what, signaller = signaller)
 
   if (is_null(cnd$with)) {
     with <- NULL
   } else {
-    with <- feature_spec(cnd$with, signaller = signaller)
+    with <- spec(cnd$with, signaller = signaller)
   }
 
   list(