+++ R/utils.R
@@ -1,7 +1,3 @@
-
-has_colour <- function() is_installed("crayon") && crayon::has_color()
-silver <- function(x) if (has_colour()) crayon::silver(x) else x
-
 upcase1 <- function(x) {
   substr(x, 1, 1) <- toupper(substr(x, 1, 1))
   x
@@ -14,6 +10,7 @@ paste_line <- function(...) {
   paste(chr(...), collapse = "\n")
 }
 
+# nocov start
 has_crayon <- function() {
   is_installed("crayon") && crayon::has_color()
 }
@@ -28,3 +25,9 @@ silver    <- function(x) if (has_crayon()) crayon::silver(x)    else x
 bold      <- function(x) if (has_crayon()) crayon::bold(x)      else x
 italic    <- function(x) if (has_crayon()) crayon::italic(x)    else x
 underline <- function(x) if (has_crayon()) crayon::underline(x) else x
+# nocov end
+
+lifecycle_abort <- function(x, env = parent.frame()) {
+  x <- paste0("Internal error in lifecycle: ", glue::trim(x))
+  abort(glue::glue(x, .envir = env))
+}