+++ tests/testthat/test-arg.R
@@ -1,5 +1,3 @@
-context("arg")
-
 test_that("deprecated() returns the missing argument", {
   fn <- function(foo = deprecated()) is_present(foo)
   expect_false(fn())