+++ tests/testthat/test-lifecycle.R
@@ -1,22 +1,19 @@
-context("lifecycle")
-
 test_that("deprecate_soft() warns when called from global env", {
-  old <- Sys.getenv("TESTTHAT_PKG")
-  Sys.setenv("TESTTHAT_PKG" = "")
-  on.exit(Sys.setenv("TESTTHAT_PKG" = old))
+  local_options(lifecycle_verbosity = NULL)
 
-  fn <- function(id) deprecate_soft("1.0.0", "foo()", id = id)
-  scoped_bindings(.env = global_env(), fn = fn)
+  fn <- function(id) {
+    deprecate_soft("1.0.0", "foo()", id = id)
+  }
+  expect_no_warning(fn("called from local env"))
 
-  locally(
-    expect_no_warning(fn("called from local env"))
+  local_bindings(.env = global_env(), fn = fn)
+  env_bind_lazy(
+    current_env(),
+    do = fn("called from global env"),
+    .eval_env = global_env()
   )
 
-  with_options(lifecycle_verbosity = "default", {
-    with_env(global_env(), {
-      expect_warning(fn("called from global env"), "foo")
-    })
-  })
+  expect_deprecated(do, "foo")
 })
 
 test_that("deprecate_soft() warns when called from package being tested", {
@@ -36,7 +33,7 @@ test_that("deprecate_soft() warns when option is set", {
 })
 
 test_that("deprecate_warn() repeats warnings when option is set", {
-  scoped_options(lifecycle_verbosity = "warning")
+  local_options(lifecycle_verbosity = "warning")
 
   retired1 <- function() deprecate_soft("1.0.0", "foo()", id = "signal repeat")
   retired2 <- function() deprecate_warn("1.0.0", "foo()", id = "warn repeat")
@@ -49,7 +46,7 @@ test_that("deprecate_warn() repeats warnings when option is set", {
 })
 
 test_that("can promote lifecycle warnings to errors", {
-  scoped_options(lifecycle_verbosity = "error")
+  local_options(lifecycle_verbosity = "error")
   expect_lifecycle_defunct(deprecate_soft("1.0.0", "foo()"), "was deprecated")
   expect_lifecycle_defunct(deprecate_warn("1.0.0", "foo()"), "was deprecated")
 })
@@ -60,16 +57,28 @@ test_that("soft-deprecation warnings are issued when called from child of global
 })
 
 test_that("deprecation warnings are not displayed again", {
+  local_options(lifecycle_verbosity = NULL)
+
   wrn <- catch_cnd(deprecate_warn("1.0.0", "foo()", id = "once-every-8-hours-note"))
   footer <- wrn$internal$footer
   expect_true(is_string(footer) && grepl("once every 8 hours", footer))
 
-  scoped_options(lifecycle_verbosity = "warning")
+  local_options(lifecycle_verbosity = "warning")
 
   wrn <- catch_cnd(deprecate_warn("1.0.0", "foo()", id = "once-every-8-hours-no-note"))
   expect_false(grepl("once every 8 hours", wrn$message))
 })
 
 test_that("the topenv of the empty env is not the global env", {
+  local_options(lifecycle_verbosity = NULL)
   expect_silent(deprecate_soft("1.0.0", "foo()", env = empty_env(), id = "topenv of empty env"))
+})
+
+test_that("expect_deprecated() matches regexp", {
+  expect_deprecated(deprecate_soft("1.0", "fn()", details = "foo"), "foo")
+  expect_deprecated(deprecate_warn("1.0", "fn()", details = "foo.["), "foo.[", fixed = TRUE)
+
+  expect_deprecated(expect_failure(
+    expect_deprecated(deprecate_soft("1.0", "fn()"), "foo")
+  ))
 })